Oragne 작업내용 Attendance 포함된 거(Attendance) Enrollment에 Long 수정

+++ src/main/java/com/green/muziuniv_be_notuser/entity/enrollment/Enrollment.java
@@ -22,7 +22,7 @@ public class Enrollment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int enrollmentId;
+    private Long enrollmentId;
 
     @ManyToOne
     @JoinColumn(name = "course_id", nullable = false)
