+++ src/main/java/com/green/muziuniv_be_notuser/student/enrollment/EnrollmentRepository.java
@@ -3,5 +3,5 @@ package com.green.muziuniv_be_notuser.student.enrollment;
 import com.green.muziuniv_be_notuser.entity.enrollment.Enrollment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnrollmentRepository extends JpaRepository<Enrollment, Integer> {
+public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
 }
